+++ src/main/java/net/mtrop/doom/WadFile.java
@@ -448,6 +448,7 @@ public class WadFile implements Wad, AutoCloseable
 	 * @throws IllegalArgumentException if the provided name is not a valid name.
 	 * @throws IOException if the data cannot be written.
 	 * @throws NullPointerException if <code>entryName</code> or <code>data</code> is <code>null</code>.
+	 * @since [NOW]
 	 */
 	public WadEntry addData(String entryName, byte[] data, boolean noFlush) throws IOException
 	{
@@ -482,6 +483,7 @@ public class WadFile implements Wad, AutoCloseable
 	 * @throws IllegalArgumentException if the provided name is not a valid name.
 	 * @throws IOException if the data cannot be written.
 	 * @throws NullPointerException if <code>entryName</code> or <code>data</code> is <code>null</code>.
+	 * @since [NOW]
 	 */
 	public WadEntry addDataAt(int index, String entryName, byte[] data, boolean noFlush) throws IOException
 	{
@@ -514,7 +516,7 @@ public class WadFile implements Wad, AutoCloseable
 	 * @throws IOException if the data cannot be written.
 	 * @throws ArrayIndexOutOfBoundsException if the lengths of entryNames and data do not match.
 	 * @throws NullPointerException if an object if <code>entryNames</code> or <code>data</code> is <code>null</code>.
-	 * @since NOW
+	 * @since [NOW]
 	 */
 	public WadEntry[] addAllData(String[] entryNames, byte[][] data, boolean noFlush) throws IOException
 	{
@@ -562,7 +564,7 @@ public class WadFile implements Wad, AutoCloseable
 	 * @throws IOException if the data cannot be written.
 	 * @throws ArrayIndexOutOfBoundsException if the lengths of entryNames and data do not match.
 	 * @throws NullPointerException if an object if <code>entryNames</code> or <code>data</code> is <code>null</code>.
-	 * @since NOW
+	 * @since [NOW]
 	 */
 	public WadEntry[] addAllDataAt(int index, String[] entryNames, byte[][] data, boolean noFlush) throws IOException
 	{